+++ application/app/tasks/core/task_core_debug.h
@@ -2,12 +2,13 @@
 #ifndef TASK_DEBUG_H
 #define TASK_DEBUG_H
 
-#include "debug_io.h"
-#include "system_err.h"
+#include <stdint.h>
+#include <stddef.h>
+#include <stdbool.h>
 
 void set_task_id(void *task_id);
 void task_printf( const char *pFmt, ...);
-void task_hex_dump(const char *title, const uint8_t *data, uint32_t length);
+void task_hex_dump(const char *title, const uint8_t *data, size_t length);
 void set_forced_print(bool set);
 
 #endif
